requirements.txt
pyyaml
python-dotenv
cryptography
aiofiles
sqlalchemy[async]
aiosqlite
